Adjust tests to be ran separately
