+++ package.json
@@ -26,7 +26,9 @@
     "demo-ws-contract": "node scripts/ws-contract-bribe-demo.js",
     "demo-simple": "node scripts/1559-demo.js",
     "demo-contract": "node scripts/1559-contract-bribe-demo.js",
-    "demo-private-tx": "node scripts/private-tx-demo.js"
+    "demo-private-tx": "node scripts/private-tx-demo.js",
+	"e2e-reverting-bundles": "node scripts/e2e-reverting-bundles.js",
+	"e2e-reverting-megabundle": "node scripts/e2e-reverting-megabundle.js"
   },
   "devDependencies": {
     "@types/node-fetch": "^2.5.10",
